Added .tutorder files
This makes it so that if a .tutorder files is found,
that it puts the items in the correct oder.

The problem is that I thought that by returning by reference,
that modifying a vector outside would modify the member,
but I think I was wrong. If I'm not, I'll change the function
later.